Fix bug with the JS test (#148)
All competitors technically need to update their "main.py" files if they aren't using python. However, anyone not using python is likely using the lux ai CLI tool anyway so this should not break any bots (and if any bots break, there's a simple 2 line fix)

This should also not affect any existing agents submitted to the competition as the code only breaks when you pass in a python JS agent function to the kaggle envs and not a file path.